Fixed missing column detection code.
We were using before a raw sql query "select typeof(column) from table"
which returns an empty result on the current emulator images (Lollipop).

Now we are instead using pragma table_info(table) to get the same
information. This fixes it.

However it is still untested whether this query is supported
in all api levels. If this is not the case we will need to switch
using Cursor#getType(String) and bump the minimum API version to 11.